move __DATA__ section after Pod
Data::Section understands it should stop reading at an __END__ token.
However, that is behavior build into Data::Section, not general behavior
of the *DATA handle. Since it is not generalized behavior, Pod::Weaver
will not modify content after a __DATA__ marker. Instead it will ignore
that content and do its own modifications before the marker.

Move the Pod content to before the __DATA__ marker, and omit the __END__
marker (as this would prevent perl from reading to the __DATA__ marker).
This allows Pod::Weaver to properly modify the Pod. It does mean that we
aren't purely doing modifications at the end of the file anymore, but
that would be difficult to accomplish in this case.